add unseen-random prompt picking algorithm
unseen-random will select unseen prompts by a user at random. In this
implementation, if a user has seen a prompt with a pair of choices we
will not show another prompt with the same choices in different order.
To keep things simple, the choice with the lower id is always on the
left.

We also need to account for what happens when a user is done. In those
situations, we respond with an empty prompt.